// windmill/f/actual_up/types.ts
export type UpWebhookEvent = {
  data: {
    attributes: {
      createdAt: string;
      eventType:
        | "PING"
        | "TRANSACTION_CREATED"
        | "TRANSACTION_DELETED"
        | "TRANSACTION_SETTLED";
    };
    id: string;
    relationships: {
      transaction?: { data: { id: string; type: "transactions" } | null };
      webhook: { data: { id: string; type: "webhooks" } };
    };
    type: "webhook-events";
  };
};